Feature/course service base model

+++ courses-service/src/main/java/ru/just/courses/service/CourseService.java
@@ -2,43 +2,54 @@ package ru.just.courses.service;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import ru.just.courses.controller.exception.MethodNotAllowedException;
 import ru.just.courses.dto.CourseDto;
 import ru.just.courses.dto.CreateCourseDto;
+import ru.just.courses.dto.UpdateCourseDto;
 import ru.just.courses.model.course.Course;
-import ru.just.courses.model.user.User;
 import ru.just.courses.repository.CourseRepository;
-import ru.just.courses.repository.UserRepository;
+import ru.just.securitylib.service.ThreadLocalTokenService;
 
 import java.util.List;
 import java.util.NoSuchElementException;
+import java.util.Optional;
 import java.util.stream.Collectors;
 
 @RequiredArgsConstructor
 @Service
 public class CourseService {
     private final CourseRepository courseRepository;
-    private final UserRepository userRepository;
+    private final ThreadLocalTokenService tokenService;
 
     public CourseDto saveCourse(CreateCourseDto courseDto) {
-        if (!userRepository.existsById(courseDto.getAuthorId())) {
-            throw new NoSuchElementException("user with specified authorId doesn't exists");
-        }
-        final Course course = courseRepository.save(courseDto.toEntity());
+        final Course entity = courseDto.toEntity();
+        entity.setAuthorId(tokenService.getUserId());
+        final Course course = courseRepository.save(entity);
         return new CourseDto().fromEntity(course);
     }
 
-    public void updateCourse(Long id, CreateCourseDto courseDto) {
-        Course course = courseRepository.findById(id)
-                .orElseThrow(() -> new NoSuchElementException("course with specified id doesn't exists"));
-        courseRepository.save(courseDto.toEntity().withId(course.getId()));
+    public void updateCourse(Long courseId, UpdateCourseDto courseDto) {
+        Course course = courseRepository.findById(courseId)
+                .orElseThrow(() -> new NoSuchElementException("course with specified courseId doesn't exists"));
+        if (!course.getAuthorId().equals(tokenService.getUserId())) {
+            throw new MethodNotAllowedException("You are not a course author");
+        }
+        course.setDescription(courseDto.getDescription());
+        course.setTitle(courseDto.getTitle());
+        course.setCompletionTimeInHours(courseDto.getCompletionTimeInHours());
+        courseRepository.save(course);
     }
 
     public void deleteCourseById(Long id) {
+        final Optional<Course> course = courseRepository.findById(id);
+        if (course.isPresent() && !course.get().getAuthorId().equals(tokenService.getUserId())) {
+            throw new MethodNotAllowedException("You are not a course author");
+        }
         courseRepository.deleteById(id);
     }
 
     public List<CourseDto> getCourses(String prefix) {
-        return courseRepository.findByTitleLike(prefix + "%").stream()
+        return courseRepository.findByTitleLikeIgnoreCase(prefix + "%").stream()
                 .map(new CourseDto()::fromEntity)
                 .collect(Collectors.toList());
     }
@@ -48,14 +59,5 @@ public class CourseService {
         dto.fromEntity(courseRepository.findById(id)
                 .orElseThrow(() -> new NoSuchElementException("course with specified id doesn't exists")));
         return dto;
-    }
-
-    public void assignUser(Long courseId, Long userId) {
-        User user = userRepository.findById(userId)
-                .orElseThrow(() -> new NoSuchElementException("user with specified id doesn't exists"));
-        Course course = courseRepository.findById(courseId)
-                .orElseThrow(() -> new NoSuchElementException("course with specified id doesn't exists"));
-        user.addCourse(course);
-        userRepository.save(user);
     }
 }
